fix error when /home/"${USER}"/.ssh already exists
Changes to be committed:
	modified:   entrypoint.sh

+++ entrypoint.sh
@@ -29,8 +29,13 @@ fi;
 
 # ssh keys
 if [ -d "${SSH_KEYDIR}" ]; then
-    echo "> Setting SSH key 🔑 at $SSH_KEYDIR"
-    ln -s "${SSH_KEYDIR}" /home/"${USER}"/.ssh
+    if [ ! -L /home/"${USER}"/.ssh ]
+    then
+        echo "> Setting SSH key 🔑 at $SSH_KEYDIR"
+        ln -s "${SSH_KEYDIR}" /home/"${USER}"/.ssh
+    else
+        echo "> Setting SSH key 🔑: keys already exists /home/${USER}/.ssh"
+    fi
 fi
 
 # jupyterlab-lsp
